rebase feature/27/add-store-list-update-image-sticker to origin/develop
